Cypress External Response:41 7/21/2021 2:46 AM - Hemanth R Hi Rossi Kwak,
Thank you for the files shared. We were able to check that endpoint is not active but still not able to verify if previous transfer is still active. If either of these is true, that would cause Invalid Sequence error.

Can you please check below.
1. Test the attached firmware file, where we are checking for the socket status before CyU3PUsbHostEpSetXfer() is called.
2. Let me know if any of the below APIs is called before the error occurs:
   CyU3PUsbHostEpAbort(), CyU3PUsbHostPortDisable(), CyU3PUsbHostEpRemove()

After this test, we will work on verifying if there is any old transfer pending.

Please let me know if there are any questions.

Thanks and Regards,
Hemanth